requirements.txt
kornia==0.7.3
matplotlib==3.8.0
numpy==1.24.3
Pillow==10.0.1
PyYAML==6.0.1
scikit_learn==1.3.0
scipy==1.14.0
thop==0.1.1.post2209072238
torch==2.1.0
torchvision==0.16.1
tqdm==4.65.0
wandb==0.17.4
